@@ -5,7 +5,7 @@ requires = ["poetry-core"]
 
 [tool.poetry]
 name = "gcop"
-version = "1.7.3"
+version = "2.0.0"
 description = "gcop is your git AI copilot."
 readme = "README.md"
 authors = ["Zeeland <zeeland4work@gmail.com>"]
@@ -18,14 +18,15 @@ keywords = ["gcop", "git", "git copilot", "promptulate", "pne", "LLM", "zeeland"
 "gcop" = "gcop.__main__:app"
 
 [tool.poetry.dependencies]
-python = ">=3.8.1,<4.0"
+python = ">=3.9,<4.0"
 zeeland = ">=0.3.0,<1.0.0"
 pne = "^1.1.1"
 click = "^8.1.7"
 questionary = "^2.0.1"
-typer = "^0.12.3"
-rich = "^13.7.1"
+typer = "0.4.2"
+rich = ">=10.14.0,<11.0.0"
 pyyaml = "^6.0.1"
+conftier = "^0.2.0"
 
 [tool.poetry.dev-dependencies]
 pytest = "^8.3.3"
